stock_analyzer/services/index_engine/requirements.txt
pandas==2.2.0
python-dotenv==1.0.1
psycopg[binary]
psycopg-pool
requests==2.31.0